Refactor Sidesheet to typescript.
