+++ libraries/core-react/src/SideSheet/SideSheet.tokens.ts
@@ -1,4 +1,3 @@
-// @ts-nocheck
 import { tokens } from '@equinor/eds-tokens'
 
 const {
