+++ libraries/core-react/src/SideSheet/index.ts
@@ -1,2 +1,1 @@
-// @ts-nocheck
 export { SideSheet } from './SideSheet'
